Renaming CurrentSessionView into SessionInfoView to be more generic
