+++ vector/src/main/res/layout/view_session_info.xml
@@ -8,7 +8,7 @@
     android:paddingBottom="16dp">
 
     <ImageView
-        android:id="@+id/currentSessionDeviceTypeImageView"
+        android:id="@+id/sessionInfoDeviceTypeImageView"
         android:layout_width="40dp"
         android:layout_height="40dp"
         android:layout_marginTop="16dp"
@@ -21,18 +21,18 @@
         tools:src="@drawable/ic_device_type_mobile" />
 
     <TextView
-        android:id="@+id/currentSessionNameTextView"
+        android:id="@+id/sessionInfoNameTextView"
         style="@style/TextAppearance.Vector.Subtitle.Medium.DevicesManagement"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="4dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/currentSessionDeviceTypeImageView"
+        app:layout_constraintTop_toBottomOf="@id/sessionInfoDeviceTypeImageView"
         tools:text="Element Mobile: Android" />
 
     <LinearLayout
-        android:id="@+id/currentSessionVerificationStatusContainer"
+        android:id="@+id/sessionInfoVerificationStatusContainer"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="12dp"
@@ -40,17 +40,17 @@
         android:orientation="horizontal"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/currentSessionNameTextView">
+        app:layout_constraintTop_toBottomOf="@id/sessionInfoNameTextView">
 
         <im.vector.app.core.ui.views.ShieldImageView
-            android:id="@+id/currentSessionVerificationStatusImageView"
+            android:id="@+id/sessionInfoVerificationStatusImageView"
             android:layout_width="16dp"
             android:layout_height="16dp"
             android:importantForAccessibility="no"
             tools:src="@drawable/ic_shield_trusted" />
 
         <TextView
-            android:id="@+id/currentSessionVerificationStatusTextView"
+            android:id="@+id/sessionInfoVerificationStatusTextView"
             style="@style/TextAppearance.Vector.Body"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -60,7 +60,7 @@
     </LinearLayout>
 
     <TextView
-        android:id="@+id/currentSessionVerificationStatusDetailTextView"
+        android:id="@+id/sessionInfoVerificationStatusDetailTextView"
         style="@style/TextAppearance.Vector.Body.DevicesManagement"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -69,11 +69,11 @@
         android:gravity="center"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/currentSessionVerificationStatusContainer"
+        app:layout_constraintTop_toBottomOf="@id/sessionInfoVerificationStatusContainer"
         tools:text="@string/device_manager_verification_status_detail_verified" />
 
     <Button
-        android:id="@+id/currentSessionVerifySessionButton"
+        android:id="@+id/sessionInfoVerifySessionButton"
         android:layout_width="0dp"
         android:layout_height="52dp"
         android:layout_marginHorizontal="24dp"
@@ -81,10 +81,10 @@
         android:text="@string/device_manager_verify_session"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/currentSessionVerificationStatusDetailTextView" />
+        app:layout_constraintTop_toBottomOf="@id/sessionInfoVerificationStatusDetailTextView" />
 
     <Button
-        android:id="@+id/currentSessionViewDetailsButton"
+        android:id="@+id/sessionInfoViewDetailsButton"
         style="@style/Widget.Vector.Button.Text"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -93,6 +93,6 @@
         android:text="@string/device_manager_view_details"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/currentSessionVerifySessionButton" />
+        app:layout_constraintTop_toBottomOf="@id/sessionInfoVerifySessionButton" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
